Relocate and rearrange surprisal extractors
