+++ text_metrics/tests/test_surprisal_extraction.py
@@ -1,7 +1,7 @@
 import unittest
 from text_metrics.utils import clean_text
 from text_metrics.ling_metrics_funcs import get_surprisal
-from text_metrics.surprisal_extractors import CatCtxLeftSurpExtractor
+from text_metrics.surprisal_extractors.SurprisalExtractor import CatCtxLeftSurpExtractor
 
 
 class TestSurprisalExtraction(unittest.TestCase):
